+++ src/pages/home/home.module.css
@@ -1,4 +1,4 @@
-.app {
+.home {
   display: flex;
   flex-direction: column;
   height: 100vh;
