+++ src/pages/login.module.css
@@ -1,9 +1,35 @@
 .formContainer {
-  outline: 1px solid red;
-  width: 480px;
-  min-height: 400px;
-  margin-top: 120px;
-  margin-right: auto;
-  margin-left: auto;
+  display: flex;
+  flex-direction: column;
+  padding-top: 180px;
+  justify-content: center;
+  align-items: center;
 }
+
+.form {
+  max-width: 480px;
+  outline: 1px solid green;
+}
+
+.headline {
+  text-align: center;
+}
+
+.inputItems {
+  padding: 0;
+  border: none;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.button {
+  outline: 1px solid blueviolet;
+  display: flex;
+}
+
+
+
+
